assignment/ai_ml/requirements.txt
fastapi
uvicorn
openai
langchain
unstructured
langchain-openai
langchain-community
langchain_pinecone
pypdf
pinecone-client